Move refreshFileHandlers to OpsLogger
Java 8 allows static methods on interfaces, which provides
a much nicer place to put this method.

(This does mean that users could call the method from logging
code, but it's static, so they could call it anyway).
